Fixed feral pills, broke the job worker.
