+++ Defs/Hediffs/Pawnmorph_Parts.xml
@@ -1,4 +1,3 @@
-<?xml version="1.0" encoding="utf-8" ?>
 <Defs>
 	<HediffDef Name="PawnmorphPart" Abstract="True">
 		<hediffClass>Pawnmorph.Hediff_AddedMutation</hediffClass>
