+++ Patches/BodyAddons.xml
@@ -1801,7 +1801,5 @@
 				</value>
 			</li>
 		</operations>
-
 	</Operation>
-
 </Patch>
