+++ Source/Pawnmorphs/Esoteria/Hediff_Morph.cs
@@ -2,15 +2,13 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using UnityEngine;
+using RimWorld;
+using Verse;
 
 namespace Pawnmorph
 {
-    public enum TFGender : byte
+    class Hediff_Morph : HediffWithComps
     {
-        None,
-        Male,
-        Female,
-        Switch,
-        Original
     }
 }
